Pure pursuit and pursuit test that work
